backend/requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
django-silk==5.0.4

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
django-stubs==4.2.7

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
model-bakery==1.17.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Development server
django-livereload-server==0.4